+++ index.html
@@ -1,64 +1,64 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<link
-			rel="shortcut icon"
-			type="image/jpg"
-			href="./resources/images/favIcon.png"
-		/>
-		<link rel="stylesheet" href="./styles/index.css" />
-		<link
-			rel="stylesheet"
-			href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css"
-			integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p"
-			crossorigin="anonymous"
-		/>
-		<script src="./dist/bundle.js" type="module" defer></script>
-		<title>Document</title>
-	</head>
-	<body>
-		<div class="app-wrapper">
-			<header id="js-header"></header>
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link
+      rel="shortcut icon"
+      type="image/jpg"
+      href="./resources/images/favIcon.png"
+    />
+    <link rel="stylesheet" href="./styles/index.css" />
+    <link
+      rel="stylesheet"
+      href="https://pro.fontawesome.com/releases/v5.10.0/css/all.css"
+      integrity="sha384-AYmEC3Yw5cVb3ZcuHtOA93w35dYTsvhLPVnYs9eStHfGJvOvKxVfELGroGkvsg+p"
+      crossorigin="anonymous"
+    />
+    <script src="./dist/bundle.js" type="module" defer></script>
+    <title>Document</title>
+  </head>
+  <body>
+    <div class="app-wrapper">
+      <header id="js-header"></header>
 
-			<main>
-				<div class="home-content">
-					<div class="home-content-aside" id="side-bar">
-						<!-- javascript  rendered -->
-					</div>
+      <main>
+        <div class="home-content">
+          <div class="home-content-aside" id="side-bar">
+            <!-- javascript  rendered -->
+          </div>
 
-					<div class="home-content-main" id="main-content">
-						<div class="home-content-main-img">
-							<img src="./resources/images/main-img.jpg" alt="" />
-						</div>
-						<div class="home-content-offers">
-							<div class="home-content-offers-title">
-								<h2>Mirá las ofertas de hoy!!!</h2>
-							</div>
-							<div
-								class="home-content-offers-products"
-								id="product-carrousel-container"
-							>
-								<!-- Javascript rendered -->
-							</div>
-						</div>
+          <div class="home-content-main" id="main-content">
+            <div class="home-content-main-img">
+              <img src="./resources/images/main-img.jpg" alt="" />
+            </div>
+            <div class="home-content-offers">
+              <div class="home-content-offers-title">
+                <h2>Mirá las ofertas de hoy!!!</h2>
+              </div>
+              <div
+                class="home-content-offers-products"
+                id="product-carrousel-container"
+              >
+                <!-- Javascript rendered -->
+              </div>
+            </div>
 
-						<div class="home-content-partner">
-							<a href="./signUp.html" class="home-content-partner-link">
-								<div>
-									<i class="fas fa-hands-helping fa-2x"></i>
-									<h2>Hacete socio de nuestro club de descuentos!!!</h2>
-								</div>
-								<p>Hacé click aquí para completar el formulario!!!</p>
-							</a>
-						</div>
-					</div>
-				</div>
-			</main>
+            <div class="home-content-partner">
+              <a href="./signUp.html" class="home-content-partner-link">
+                <div>
+                  <i class="fas fa-hands-helping fa-2x"></i>
+                  <h2>Hacete socio de nuestro club de descuentos!!!</h2>
+                </div>
+                <p>Hacé click aquí para completar el formulario!!!</p>
+              </a>
+            </div>
+          </div>
+        </div>
+      </main>
 
-			<footer id="js-footer"></footer>
-		</div>
-	</body>
+      <footer id="js-footer"></footer>
+    </div>
+  </body>
 </html>
