--- src/CartItem.ts
@@ -1,6 +1,0 @@
-import Product from "./Product";
-
-export default interface CartItem {
-	item: Product;
-	amount: number;
-}
