--- src/Comment.ts
@@ -1,7 +1,0 @@
-export default interface Comment {
-	postId: number;
-	id: number;
-	name: string;
-	email: string;
-	body: string;
-}
