--- src/Path.ts
@@ -1,4 +1,0 @@
-export default interface Path {
-	path: string;
-	viewRendering: Function;
-}
