+++ src/components/product.ts
@@ -1,7 +1,7 @@
-import Product from "../Product";
+import Product from "../classes/Product";
 
 export default function productComponent(product: Product): string {
-	return `
+  return `
     <a href="./productDetails.html?id=${product.id}" class="content-product">
         <img src="${product.image}" alt="${product.name}" />
         <p>${product.name}</p>
