+++ src/components/productCarrousel.ts
@@ -1,10 +1,10 @@
-import Product from "../Product";
+import Product from "../classes/Product";
 
 export default function productCarrousel(product: Product): string {
-	return `
+  return `
     <a href="./productDetails.html?id=${
-			product.id
-		}" class="home-content-offers-product">
+      product.id
+    }" class="home-content-offers-product">
         <img src="${product.thumbnail}" alt="${product.name}" />
         <p>${product.name}</p>
         <h2>$${(product.price * (1 - product.discount / 100)).toFixed(0)}</h2>
