+++ src/components/productComments.ts
@@ -1,8 +1,8 @@
-import Comment from "../Comment";
+import Comment from "../interfaces/CommentInterface";
 
 export default function productComments(comment: Comment): string {
-	let name = comment.email.split("@");
-	return `
+  let name = comment.email.split("@");
+  return `
     <div class="product-details-comments-item">
         <h3>Name: ${name[0]}</h3>
         <p>
