+++ src/components/userComponent.ts
@@ -1,9 +1,9 @@
-import Store from "../Store";
+import Store from "../classes/Store";
 
 export default function userComponent(store: Store) {
-	let returnComponent: string;
-	if (store.user) {
-		returnComponent = `
+  let returnComponent: string;
+  if (store.user) {
+    returnComponent = `
 		<h2>Nombre: ${store.user.firstName.toUpperCase()} </h2>
 		<h2>Apellido: ${store.user.lastName.toUpperCase()} </h2>
 		<h2>Email: ${store.user.mailAdress.toUpperCase()} </h2>
@@ -14,10 +14,10 @@ export default function userComponent(store: Store) {
 		</select>
 		<button class="button-link" id="log-out">LOG-OUT</button>
 		`;
-	} else {
-		returnComponent = `
+  } else {
+    returnComponent = `
 		<h2>USUARIO INVITADO </h2>
 		`;
-	}
-	return returnComponent;
+  }
+  return returnComponent;
 }
