+++ src/interfaces/CartItemInterface.ts
@@ -1,0 +1,6 @@
+import Product from "../classes/Product";
+
+export default interface CartItemInterface {
+  item: Product;
+  amount: number;
+}
