+++ src/interfaces/CommentInterface.ts
@@ -1,0 +1,7 @@
+export default interface CommentInterface {
+  postId: number;
+  id: number;
+  name: string;
+  email: string;
+  body: string;
+}
