+++ src/interfaces/PathInterface.ts
@@ -1,0 +1,4 @@
+export default interface PathInterface {
+  path: string;
+  viewRendering: Function;
+}
