+++ src/view-logic/index.ts
@@ -1,0 +1,13 @@
+import productCarrousel from "../components/productCarrousel";
+import Product from "../classes/Product";
+import Store from "../classes/Store";
+
+export default function returnIndexView(store: Store) {
+  let productCarrouselContainer: HTMLElement = document.getElementById(
+    "product-carrousel-container"
+  )!;
+
+  store.showCatalog().forEach((product: Product) => {
+    productCarrouselContainer.innerHTML += productCarrousel(product);
+  });
+}
